Refactoring monitoring window specific function to corresponding engine
