+++ python/hsfs/core/monitoring_window_config_engine.py
@@ -13,14 +13,14 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 #
-from typing import Optional, Union, Tuple
+from typing import List, Optional, TypeVar, Union, Tuple
 import re
 from datetime import datetime, timedelta
 
 from hsfs.core import monitoring_window_config as mwc
 from hsfs.core.feature_descriptive_statistics import FeatureDescriptiveStatistics
 from hsfs.util import convert_event_time_to_timestamp
-from hsfs import feature_group, feature_view
+from hsfs import feature_group, feature_view, engine
 from hsfs.core import statistics_engine
 
 
@@ -30,6 +30,12 @@ class MonitoringWindowConfigEngine:
     def __init__(self) -> "MonitoringWindowConfigEngine":
         # No need to initialize anything
         pass
+
+    def _init_statistics_engine(self, feature_store_id: int, entity_type: str):
+        self._statistics_engine = statistics_engine.StatisticsEngine(
+            feature_store_id=feature_store_id,
+            entity_type=entity_type,
+        )
 
     def validate_monitoring_window_config(
         self,
@@ -243,3 +249,193 @@ class MonitoringWindowConfigEngine:
             convert_event_time_to_timestamp(start_time),
             convert_event_time_to_timestamp(end_time),
         )
+
+    def run_single_window_monitoring(
+        self,
+        entity: Union["feature_group.FeatureGroup", "feature_view.FeatureView"],
+        monitoring_window_config: "mwc.MonitoringWindowConfig",
+        feature_name: Optional[str] = None,
+        use_event_time: bool = False,
+        transformation_function_dataset_version=None,
+    ) -> Union[FeatureDescriptiveStatistics, List[FeatureDescriptiveStatistics]]:
+        """Fetch the entity data based on monitoring window configuration and compute statistics.
+
+        Args:
+            entity: FeatureStore: Feature store to fetch the entity to monitor.
+            monitoring_window_config: MonitoringWindowConfig: Monitoring window config.
+            feature_name: str: Name of the feature to monitor.
+            use_event_time: bool: Whether to use event time or ingestion time.
+                Feature View only. Defaults to False.
+            transformation_function_dataset_version: int: Version of the dataset to use for transformation function.
+                Feature View only. Defaults to None, meaning no transformation function are applied.
+
+        Returns:
+            [FeatureDescriptiveStatistics, List[FeatureDescriptiveStatitics]]: List of Descriptive statistics.
+        """
+        self._init_statistics_engine(entity.feature_store_id, entity.ENTITY_TYPE)
+        # Check if statistics already exists
+        (start_time, end_time,) = self.get_window_start_end_times(
+            monitoring_window_config=monitoring_window_config,
+        )
+        registered_stats = self._statistics_engine.get_by_commit_time_window(
+            entity,
+            start_time=start_time,
+            end_time=end_time,
+            feature_name=feature_name,
+            row_percentage=monitoring_window_config.row_percentage,
+        )
+
+        if registered_stats is None:  # if statistics don't exist
+            # Fetch the actual data for which to compute statistics based on row_percentage and time window
+            entity_feature_df = self.fetch_entity_data_in_monitoring_window(
+                entity=entity,
+                feature_name=feature_name,
+                start_time=start_time,
+                end_time=end_time,
+                row_percentage=monitoring_window_config.row_percentage,
+                use_event_time=use_event_time,
+                transformation_function_dataset_version=transformation_function_dataset_version,
+            )
+
+            # Compute statistics on the feature dataframe
+            registered_stats = (
+                self._statistics_engine.compute_and_save_monitoring_statistics(
+                    entity,
+                    feature_dataframe=entity_feature_df,
+                    start_time=start_time,
+                    end_time=end_time,
+                    row_percentage=monitoring_window_config.row_percentage,
+                    feature_name=feature_name,
+                )
+            )
+
+        return (
+            registered_stats.feature_descriptive_statistics[0]
+            if feature_name is not None
+            else registered_stats.feature_descriptive_statistics
+        )
+
+    def fetch_entity_data_in_monitoring_window(
+        self,
+        entity: Union["feature_group.FeatureGroup", "feature_view.FeatureView"],
+        start_time: int,
+        end_time: int,
+        row_percentage: float,
+        feature_name: Optional[str] = None,
+        use_event_time: bool = False,
+        transformation_function_dataset_version: Optional[int] = None,
+    ) -> TypeVar("pyspark.sql.DataFrame"):
+        """Fetch the entity data based on time window and row percentage.
+
+        Args:
+            entity: Union[FeatureGroup, FeatureView]: Entity to monitor.
+            feature_name: str: Name of the feature to monitor.
+            start_time: int: Window start commit or event time
+            end_time: int: Window end commit or event time
+            row_percentage: fraction of rows to include [0, 1.0]
+            use_event_time: bool: Whether to use event time or ingestion time.
+                Feature View only.
+            transformation_function_dataset_version: int: Version of the dataset
+                to fetch statistics from for the transformation function.
+                Feature View only.
+
+        Returns:
+            `pyspark.sql.DataFrame`. A Spark DataFrame with the entity data
+        """
+        if entity.ENTITY_TYPE == "featuregroups":
+            # use_event_time and transformation_function don't apply here
+            entity_df = self.fetch_feature_group_data(
+                entity=entity,
+                feature_name=feature_name,
+                start_time=start_time,
+                end_time=end_time,
+            )
+
+        elif entity.ENTITY_TYPE == "featureview":
+            entity_df = self.fetch_feature_view_data(
+                entity=entity,
+                feature_name=feature_name,
+                start_time=start_time,
+                end_time=end_time,
+                use_event_time=use_event_time,
+                transformation_function_dataset_version=transformation_function_dataset_version,
+            )
+
+        if row_percentage < 1.0:
+            entity_df = entity_df.sample(fraction=row_percentage)
+
+        return entity_df
+
+    def fetch_feature_view_data(
+        self,
+        entity: "feature_view.FeatureView",
+        feature_name: Optional[str] = None,
+        start_time: Optional[int] = None,
+        end_time: Optional[int] = None,
+        use_event_time: bool = False,
+        transformation_function_dataset_version: Optional[int] = None,
+    ) -> TypeVar("pyspark.sql.DataFrame"):
+        """Fetch the feature view data based on time window and row percentage.
+
+        Args:
+            entity: FeatureView: Feature view to monitor.
+            feature_name: str: Name of the feature to monitor.
+            start_time: int: Window start commit or event time.
+            end_time: int: Window end commit or event time.
+            use_event_time: bool: Whether to use event time or ingestion time.
+            transformation_function_dataset_version: int: Dataset version of the transformation function
+
+        Returns:
+            `pyspark.sql.DataFrame`. A Spark DataFrame with the entity data
+        """
+        if use_event_time:
+            entity_df = entity._feature_view_engine.get_batch_query(
+                feature_view_obj=entity,
+                start_time=start_time,
+                end_time=end_time,
+                with_label=True if feature_name is None else False,
+                training_dataset_version=transformation_function_dataset_version,
+            ).read()
+        else:
+            entity_df = entity.query.as_of(
+                exclude_until=start_time, wallclock_time=end_time
+            ).read()
+
+        if feature_name:
+            entity_df = entity_df.select(feature_name)
+
+        if transformation_function_dataset_version:
+            entity.init_batch_scoring(
+                training_dataset_version=transformation_function_dataset_version
+            )
+
+            return engine.get_instance()._apply_transformation_function(
+                entity._batch_scoring_server._get_transformation_fns(),
+                dataset=entity_df,
+            )
+        else:
+            return entity_df
+
+    def fetch_feature_group_data(
+        self,
+        entity: "feature_group.FeatureGroup",
+        feature_name: Optional[str] = None,
+        start_time: Optional[int] = None,
+        end_time: Optional[int] = None,
+    ) -> TypeVar("pyspark.sql.Dataframe"):
+        """Fetch the feature group data based on time window.
+
+        Args:
+            entity: FeatureGroup: Feature group to monitor.
+            feature_name: str: Name of the feature to monitor.
+            start_time: int: Window start commit time.
+            end_time: int: Window end commit time.
+        """
+        if feature_name:
+            pre_df = entity.select(features=[feature_name])
+        else:
+            pre_df = entity
+
+        full_df = pre_df.as_of(exclude_until=start_time, wallclock_time=end_time).read()
+
+        return full_df
